pyproject.toml
[tool.poetry]
name = "automate-email"
version = "0.1.0"
description = ""
authors = ["Alok Desai <alokdesai2004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mime = "^0.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
